Extract BLProvider to its own file
