+++ spec/alexandria/book_providers/bl_provider_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+# This file is part of Alexandria.
+#
+# See the file README.md for authorship and licensing information.
+
+require "spec_helper"
+
+describe Alexandria::BookProviders::BLProvider do
+  it "works" do
+    assert_correct_search_result(described_class, "9781853260803")
+  end
+end
